requirements.txt
gunicorn
selenium==3.141.0
schedule
moviepy
Pillow
pygame
undetected-chromedriver